requirements.txt
Flask==2.2.3
Pillow==9.6.0
